rename cmath.snap => cmath.align
